+++ pkg/apis/deployment/v1/external_access_type.go
@@ -38,12 +38,15 @@ const (
 	ExternalAccessTypeLoadBalancer ExternalAccessType = "LoadBalancer"
 	// ExternalAccessTypeNodePort yields a cluster with a service of type `NodePort` to provide external access
 	ExternalAccessTypeNodePort ExternalAccessType = "NodePort"
+	// ExternalAccessTypeManaged yields a cluster with a service which controls only selector.
+	ExternalAccessTypeManaged ExternalAccessType = "Managed"
 )
 
 func (t ExternalAccessType) IsNone() bool         { return t == ExternalAccessTypeNone }
 func (t ExternalAccessType) IsAuto() bool         { return t == ExternalAccessTypeAuto }
 func (t ExternalAccessType) IsLoadBalancer() bool { return t == ExternalAccessTypeLoadBalancer }
 func (t ExternalAccessType) IsNodePort() bool     { return t == ExternalAccessTypeNodePort }
+func (t ExternalAccessType) IsManaged() bool      { return t == ExternalAccessTypeManaged }
 
 // AsServiceType returns the k8s ServiceType for this ExternalAccessType.
 // If type is "Auto", ServiceTypeLoadBalancer is returned.
@@ -62,7 +65,8 @@ func (t ExternalAccessType) AsServiceType() core.ServiceType {
 // Return errors when validation fails, nil on success.
 func (t ExternalAccessType) Validate() error {
 	switch t {
-	case ExternalAccessTypeNone, ExternalAccessTypeAuto, ExternalAccessTypeLoadBalancer, ExternalAccessTypeNodePort:
+	case ExternalAccessTypeNone, ExternalAccessTypeAuto, ExternalAccessTypeLoadBalancer, ExternalAccessTypeNodePort,
+		ExternalAccessTypeManaged:
 		return nil
 	default:
 		return errors.WithStack(errors.Wrapf(ValidationError, "Unknown external access type: '%s'", string(t)))