+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - (Refactor) Use cached member's clients
 - (Feature) Move PVC resize action to high-priority plan
 - (Feature) Remove forgotten ArangoDB jobs during restart
+- (Feature) Add support for managed services
 
 ## [1.2.14](https://github.com/arangodb/kube-arangodb/tree/1.2.14) (2022-07-14)
 - (Feature) Add ArangoSync TLS based rotation